move checker under pkg/ , the convention is so
